refactor: renaming export-methods package + moving domain checking
