+++ src/scripts/events/exportAll.bg.js
@@ -1,6 +1,6 @@
 import {clickOnListElt} from "../units/interface/interact-DOM/interact.tab";
 
-import {launchExport} from "../process/scraper/launchScrapping";
+import {launchScrapping} from "../process/scraper/launchScrapping.tab";
 
 export function exportAllThreadsListener() {
   let currentIndex = 0;
@@ -71,7 +71,7 @@ export function exportAllThreadsListener() {
 export function scrapOnLoadListener(domain) {
   chrome.runtime.onMessage.addListener(async function (request, sender, sendResponse) {
     if (request.message === 'executeScript') {
-      if (request.index > 0) launchExport(domain);
+      if (request.index > 0) launchScrapping(domain);
       clickOnListElt(request.index)
       setTimeout(function () {
         sendResponse({message: 'scriptExecuted'});
