+++ src/scripts/process/scraper/checkDomain.tab.js
@@ -1,0 +1,13 @@
+export const LOAD_DOMAINS = {
+  "Phind": "www.phind.com",
+  "Perplexity": "www.perplexity.ai",
+  "MaxAIGoogle": "www.google.com",
+};
+
+export const EXPORT_DOMAINS = {
+  "PhindSearch": "www.phind.com/search",
+  "PhindChat": "www.phind.com/agent",
+  "Perplexity": "www.perplexity.ai/search",
+  "MaxAIGoogle": "www.google.com/search",
+};
+
