+++ src/scripts/process/tab.js
@@ -1,12 +1,10 @@
 // import infos from "./infos";
-import {
-  domainLoadChecker
-} from "../units/processing/checker/domainChecker.all";
 import {checkClickCountAndDisplayModal} from "../units/interface/modals/clickCount.all";
 import {uiEnhancer} from "../events/uiEnhancer.tab";
 import {detectPageLoad} from "../events/detectPageLoad.tab";
-import {launchExport} from "./scraper/launchScrapping";
-import {domainExportChecker} from "./scraper/steps/checkDomain.tab";
+import {launchScrapping} from "./scraper/launchScrapping.tab";
+import {EXPORT_DOMAINS, LOAD_DOMAINS} from "./scraper/checkDomain.tab";
+import {domainChecker} from "../units/processing/checker/domainChecker.all";
 
 async function tab() {
   // console.log(infos.APP_MODE)
@@ -16,20 +14,20 @@ async function tab() {
   });
 }
 
-export async function actionExtensionIconClicked() {
-  const domainPage = await domainExportChecker();
-  if (domainPage === null) return;
-  launchExport(domainPage);
-  checkClickCountAndDisplayModal(domainPage);
-}
-
 export async function actionPageLoaded() {
-  const domain = await domainLoadChecker();
+  const domain = await domainChecker(LOAD_DOMAINS);
   if (domain === null) return;
   const htmlCheck = detectPageLoad(domain);
   if (!htmlCheck) return;
   // scrapOnLoadListener();
   uiEnhancer(domain);
+}
+
+export async function actionExtensionIconClicked() {
+  const domainPage = await domainChecker(EXPORT_DOMAINS);
+  if (domainPage === null) return;
+  launchScrapping(domainPage);
+  checkClickCountAndDisplayModal(domainPage);
 }
 
 tab();
