+++ src/scripts/units/utils/getters.tab.js
@@ -17,3 +17,8 @@ export function getHostAndPath() {
 export function isHomepageCheck() {
   return window.location.href === "https://www.phind.com" || window.location.href === "https://www.phind.com/"
 }
+
+function hasSubpages(urlString) {
+  const regex = /^(https?|ftp|ssh|mailto):\/\/[a-z0-9:%_+.,#?!@&=-]+\/?$/;
+  return !regex.test(urlString);
+}
